Add wrapper components and move some files
