+++ src/panels/resources-panel/components/human-resources.css
