+++ src/panels/sell-panel/components/sell-pricechart.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import Chart from "react-apexcharts";
-import { formatTimeWithoutParenthesis } from "../../shared/utils/format-time";
+import { formatTimeWithoutParenthesis } from "../../../shared/utils/format-time";
 
 const getCurrentTimeFormattable = () => {
   const currentTime = new Date();
